💄 chore(docs-kit): move style files out of routes dir and exempt framework default exports
- routes/ now holds only route modules; homeStyle -> components/Home, notFoundStyle -> styles
- eslint: allow default exports where React Router framework mode requires them
- add approved home-page configurability spec
