+++ eslint.config.mjs
@@ -53,4 +53,19 @@ export default defineConfig(
       'react-refresh/only-export-components': 'off',
     },
   },
+  {
+    // React Router framework mode requires default exports from route modules,
+    // root, server entry, and the route config; the .d.ts mirrors a virtual
+    // module's default export shape.
+    files: [
+      'packages/docs-kit/site/app/routes/*.tsx',
+      'packages/docs-kit/site/root.tsx',
+      'packages/docs-kit/site/entry.server.tsx',
+      'packages/docs-kit/site/routes.ts',
+      'packages/docs-kit/site/types/*.d.ts',
+    ],
+    rules: {
+      'no-restricted-syntax': 'off',
+    },
+  },
 );
